requirements.txt
ipython
mmwebbot
python-decouple
lxml
cssselect
